Scald Core: Prefix the global variable used in a callback
Previously, a global variable named $override was created to
communicate with the callback used in scald_sas_to_rendered,
which could cause a collision if another module used it for
something else. The variable was renamed to $scald_override
to prevent this from happening.

A nicer solution would be to use the anonymous functions introduced
in PHP 5.3, and their lovely ability to inherit a specific vraiable
from the parent scope, but working with PHP 5.2 is still quite
important