Scald Core: Deals with not yet configured transcoder when unregistering
If a new type is registered, but the user doesn't save the context
form to associate it with a transcoder in all the contexts, then
the type might not have a transcoder set.

This was problematic in scald_unregister_atom, which tries to notifiy
all the configured transcoders of the atom unregistration. The code
now checks if a transcoder has been specified, and just skips the
context is no transcoder were set.