Scald Core: Move author to taxonomy terms
Previously, Scald Authors were stored in a custom table,
and the association between atoms and authors was managed
fully by Scald, which needed to provide its own CRUD API
to create, delete and search for an author. This also
means that the properties of an atom were all defined by
Scald.

This commits moves the Authors to taxonomy terms, creating
a new bundle for them when Scald is enabled, and putting
fields on them to match the authors that were in the D6
version.

Scald Quickstart creates Taxonomy term references fields,
configured with this new bundle, on all the Atom types.