Scald DnD Library: Check the correct permission for the library
The view was still configured to check the "use scald" permission,
which doesn't exist anymore in D7. Thus, only administators could
access the library, that's no longer the case.

Now that anyone can access the library, we also check if they're
allowed to create an atom of this type before showing the quick
add button.