Scald Dailymotion: Open "Search on DailyMotion" in a new window.
This is a 'hotfix' to prevent users clicking on the "Search on Daily
Motion" link in the modal window from clobbering all the values they
previously enterred, and having to start from scratch.

The long term fix we're working on is to ajax-ify this form,
re-integrating the search form in the atom creation process.