DnD: Remove bundled BeautyTips library, add optional qTip integration
Previously, user could get an atom preview by clicking on an element
in the library, which opened a preview tooltip powered by an embedded
version of the BeautyTips library.

The embedded version now got removed, and replaced by a thin
compatibility layer with the qTip module and library. Users willing to
get the preview now needs to either download and enable the qTip module
or ensure by themselves that the qTip jQuery library is loaded.