Scald Image: Don't hide the legend when using the figure player
Things aren't really ready for it yet. Ideally, for 1.1, the "Edit atom
properties" dialog in the CKEditor plugin should edit the <figcaption />
of the player, and update the HTML comment accordingly to let the figure
player render it correctly.

For now, let's just keep the standard user editable legend in addition
to the fixed figcaption.